requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
boto3==1.26.34
botocore==1.29.34
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Flask==2.2.2
frozenlist==1.3.3
future==0.18.2
idna==3.4
importlib-metadata==5.1.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
msrest==0.7.1
multidict==6.0.2
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
python-dateutil==2.8.2
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.11.0
azure-storage-blob==12.14.1
opencv-python==4.6.0.66
numpy==1.22.4
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tornado @ file:///opt/concourse/worker/volumes/live/2c1a63a2-006b-48ee-56b9-0cfe8b4927f9/volume/tornado_1606942321278/work
tqdm @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_6cb3db64-b7e1-4db2-ab0b-f944920e5367lekntvc5/croots/recipe/tqdm_1650891088614/work